Added input checks
As the root node is a special case where the dataset should be named /data instead of //data this check makes sure the first occurence happens. The second check is the update the dataset name if its provided with the proper path. i.e A.create_dataset("data") would be "/A/data" instead of "data" within the object store.