S3 Groups (#28)
* Tests for S3 groups

Based on the Ceph Group tests, changes as s3 uses client instead of iocxt.

* Skeleton of classes and functions

An outline of all the functions that need to be implemented.

* Added creation of Root Group on connect

* Create, Get and Del implemented

These methods are accessed through the root_group

* implemented create_absolute_path

* Added _create_group_object method

An object group can now be created. Passed test.

* Updated test

Test relied on create_groups for checking has_group_objects, this requirment has been removed and changed to create_group_object inline with the level the has_group_object is checking.

* _has_group_object method

Implemented method, checks that the body of the object has the correct signature.

* Updated test

Test relied on create_group for checking get_group_object, this requirment has been removed and changed to create_group_object inline with the level the get_group_object is checking

* get_group_object method

Passed tests

* Updated test

Test relied on create_group for checking del_group_object, this requirment has been removed and changed to create_group_object inline with the level the del_group_object is checking

* Added _del_group_object method

passed test

* Added get_links method

Similar to Ceph's implementation however the data is got via the boto3 implementation

* Using head_object instead of get_object

head_object is being used as the body of the object is not required thus this reduces the overall data footprint size of the method calls.

* Added set_attrs

To update metadata the object has to be copied, this function states to copy only if no changes have been made and to replace the current metadata with the new metadata.

* Added _group_group_link

Adds link information to the object metadata.

* Added create_group

For creation of groups the connection name has been changed to bucket_name. Otherwise, the same method as the Ceph backend was implemented. It passed all tests.

* Changed CephLink to S3Link

* Added get_group

This is the same implementation as CephGroup.

* Added del_group_link

* Added del_group

Same implementation to CephGroup.

* Updated Test

Added decode to signature so it is returned as a string.

* Added Create_Dataset within group

Passes Tests

* Added has_dataset

Passed Tests

* Implemented gets for datasets

Passed Tests

* Updated assert raises to be S3 specific

* Added deletion of dataset functionality

Passes Tests

* Added dataset link create/del methods

Passes tests